fcm_extractor/requirements.txt
# Core requirements for FCM extraction pipeline
numpy
pandas
matplotlib
plotly 
google-generativeai
python-dotenv
langsmith
langchain-core
langchain-openai
langchain-google-genai
tiktoken

# Clustering & Visualization
scikit-learn
umap-learn
hdbscan
sentence-transformers

# Document processing
python-docx

# Web visualization
networkx 